fit_multispectra working (with lambda binning); debug fitworkspace for the 3D W case.

+++ spectractor/fit/fitter.py
@@ -680,7 +680,7 @@ class FitWorkspace:
                     W = [1 / (self.data_cov[k] + model_err * model_err) for k in range(K)]
                 else:
                     W = self.W
-                res = [model[k] - self.data[k] for k in range(K)]
+                res = [model[k] - self.data[k].astype(float) for k in range(K)]
                 chisq = np.sum([res[k] @ (W[k].astype(float) * res[k]) for k in range(K)])
             elif self.W[0].ndim == 2:
                 K = len(self.W)
@@ -690,8 +690,8 @@ class FitWorkspace:
                     W = [L[k].T @ L[k] for k in range(K)]
                 else:
                     W = self.W
-                res = [model[k] - self.data[k] for k in range(K)]
-                chisq = np.sum([res[k] @ W[k] @ res[k] for k in range(K)])
+                res = [model[k] - self.data[k].astype(float) for k in range(K)]
+                chisq = np.sum([res[k] @ W[k].astype(float) @ res[k] for k in range(K)])
             else:
                 raise ValueError(f"First element of fitworkspace.W has no ndim attribute or has a dimension above 2. "
                                  f"I get W[0]={self.W[0]}")
@@ -936,7 +936,7 @@ def gradient_descent(fit_workspace, params, epsilon, niter=10, fixed_params=None
         if isinstance(fit_workspace.W, np.ndarray) and fit_workspace.W.dtype != np.object:
             residuals = (tmp_model - fit_workspace.data).flatten()
         elif isinstance(fit_workspace.W, np.ndarray) and fit_workspace.W.dtype == np.object:
-            residuals = [(tmp_model[k] - fit_workspace.data[k]) for k in range(len(fit_workspace.W))]
+            residuals = [(tmp_model[k] - fit_workspace.data[k].astype(float)) for k in range(len(fit_workspace.W))]
         else:
             raise TypeError(f"Type of fit_workspace.W is {type(fit_workspace.W)}. It must be a np.ndarray.")
         # Jacobian
@@ -970,7 +970,8 @@ def gradient_descent(fit_workspace, params, epsilon, niter=10, fixed_params=None
                 # warning ! here the data arrays indexed by k can have different lengths because outliers
                 # because W inverse covariance is block diagonal and blocks can have different sizes
                 # the philosophy is to temporarily flatten the data arrays
-                JT_W = [np.concatenate([J[ip][k].T @ W[k] for k in range(fit_workspace.W.shape[0])]).ravel()
+                JT_W = [np.concatenate([J[ip][k].T @ W[k].astype(float)
+                                        for k in range(fit_workspace.W.shape[0])]).ravel()
                         for ip in range(len(J))]
                 JT_W_J = np.array([[JT_W[ip2] @ np.concatenate(J[ip1][:]).ravel() for ip1 in range(len(J))]
                                    for ip2 in range(len(J))])
@@ -982,7 +983,6 @@ def gradient_descent(fit_workspace, params, epsilon, niter=10, fixed_params=None
         if fit_workspace.W.dtype != np.object:
             JT_W_R0 = JT_W @ residuals
         else:
-            # JT_W_R0 = np.sum([JT_W[k] @ residuals[k] for k in range(fit_workspace.data_cov.shape[0])], axis=0)
             JT_W_R0 = JT_W @ np.concatenate(residuals).ravel()
         dparams = - inv_JT_W_J @ JT_W_R0
 
@@ -996,7 +996,6 @@ def gradient_descent(fit_workspace, params, epsilon, niter=10, fixed_params=None
                     if pp > fit_workspace.bounds[ipp][1]:
                         tmp_params_2[ipp] = fit_workspace.bounds[ipp][1]
                 return fit_workspace.chisq(tmp_params_2)
-
             # tol parameter acts on alpha (not func)
             alpha_min, fval, iter, funcalls = optimize.brent(line_search, full_output=True, tol=5e-1, brack=(0, 1))
         else:
@@ -1004,6 +1003,7 @@ def gradient_descent(fit_workspace, params, epsilon, niter=10, fixed_params=None
             fval = np.copy(cost)
             funcalls = 0
             iter = 0
+
         tmp_params[ipar] += alpha_min * dparams
         # check bounds
         for ip, p in enumerate(tmp_params):
@@ -1011,6 +1011,7 @@ def gradient_descent(fit_workspace, params, epsilon, niter=10, fixed_params=None
                 tmp_params[ip] = fit_workspace.bounds[ip][0]
             if p > fit_workspace.bounds[ip][1]:
                 tmp_params[ip] = fit_workspace.bounds[ip][1]
+
         # prepare outputs
         costs.append(fval)
         params_table.append(np.copy(tmp_params))
@@ -1365,7 +1366,7 @@ def run_minimisation_sigma_clipping(fit_workspace, method="newton", epsilon=None
             my_logger.info(f"\n\tSigma-clipping step {step}/{niter_clip} (sigma={sigma_clip})")
         run_minimisation(fit_workspace, method=method, epsilon=epsilon, fix=fix, xtol=xtol, ftol=ftol, niter=niter)
         # remove outliers
-        indices_no_nan = ~np.isnan(fit_workspace.data.flatten())
+        indices_no_nan = ~np.isnan(fit_workspace.data.flatten().astype(float))
         residuals = np.abs(fit_workspace.model.flatten()[indices_no_nan]
                            - fit_workspace.data.flatten()[indices_no_nan]) / fit_workspace.err.flatten()[indices_no_nan]
         outliers = residuals > sigma_clip
